requirements.txt
PyPDF2
pyOpenSSL
reportlab

